lambda/requirements.txt
pylint
pytest
pytest-cov
pandas
python-dotenv
sqlalchemy
boto3